backend/requirements.txt
bcrypt==4.0.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.3.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1



